Delete VM when preempted
Without this, when preemptible VM is prempted and leaves the
instance in shutdown state with disk still incurring cost.

To use this flag I also need to switch to spot VM flags, which are
cost the same and have some more features. I've picked 3d max
runtime as that's already used in the script as "max workflow
runtime".